Factor common code of QgsMapToolAddCircle/Ellipse/Rectangle/RegularRectangle
Those classes have identical implementations for a number of methods.
Add a QgsMapToolAddAbstract class for that, and make them derive for it.
