+++ src/app/qgsmaptooladdcircle.cpp
@@ -25,65 +25,9 @@
 #include "qgssnapindicator.h"
 
 QgsMapToolAddCircle::QgsMapToolAddCircle( QgsMapToolCapture *parentTool, QgsMapCanvas *canvas, CaptureMode mode )
-  : QgsMapToolCapture( canvas, QgisApp::instance()->cadDockWidget(), mode )
-  , mParentTool( parentTool )
-  , mSnapIndicator( std::make_unique< QgsSnapIndicator>( canvas ) )
+  : QgsMapToolAddAbstract( parentTool, canvas, mode )
 {
   mToolName = tr( "Add circle" );
-
-  clean();
-  connect( QgisApp::instance(), &QgisApp::newProject, this, &QgsMapToolAddCircle::stopCapturing );
-  connect( QgisApp::instance(), &QgisApp::projectRead, this, &QgsMapToolAddCircle::stopCapturing );
-}
-
-QgsMapToolAddCircle::~QgsMapToolAddCircle()
-{
-  clean();
-}
-
-void QgsMapToolAddCircle::keyPressEvent( QKeyEvent *e )
-{
-  if ( e && e->isAutoRepeat() )
-  {
-    return;
-  }
-
-  if ( e && e->key() == Qt::Key_Escape )
-  {
-    clean();
-    if ( mParentTool )
-      mParentTool->keyPressEvent( e );
-  }
-
-  if ( e && e->key() == Qt::Key_Backspace )
-  {
-    if ( mPoints.size() == 1 )
-    {
-
-      if ( mTempRubberBand )
-      {
-        delete mTempRubberBand;
-        mTempRubberBand = nullptr;
-      }
-
-      mPoints.clear();
-    }
-    else if ( mPoints.size() > 1 )
-    {
-      mPoints.removeLast();
-
-    }
-    if ( mParentTool )
-      mParentTool->keyPressEvent( e );
-  }
-}
-
-void QgsMapToolAddCircle::keyReleaseEvent( QKeyEvent *e )
-{
-  if ( e && e->isAutoRepeat() )
-  {
-    return;
-  }
 }
 
 void QgsMapToolAddCircle::deactivate()
@@ -114,40 +58,8 @@ void QgsMapToolAddCircle::deactivate()
   QgsMapToolCapture::deactivate();
 }
 
-void QgsMapToolAddCircle::activate()
-{
-  clean();
-  QgsMapToolCapture::activate();
-}
-
 void QgsMapToolAddCircle::clean()
 {
-  if ( mTempRubberBand )
-  {
-    delete mTempRubberBand;
-    mTempRubberBand = nullptr;
-  }
-
-  mPoints.clear();
-
-  if ( mParentTool )
-  {
-    mParentTool->deleteTempRubberBand();
-  }
-
+  QgsMapToolAddAbstract::clean();
   mCircle = QgsCircle();
-
-  QgsVectorLayer *vLayer = static_cast<QgsVectorLayer *>( QgisApp::instance()->activeLayer() );
-  if ( vLayer )
-    mLayerType = vLayer->geometryType();
-}
-
-void QgsMapToolAddCircle::release( QgsMapMouseEvent *e )
-{
-  deactivate();
-  if ( mParentTool )
-  {
-    mParentTool->canvasReleaseEvent( e );
-  }
-  activate();
 }
