+++ src/app/qgsmaptooladdellipse.cpp
@@ -14,8 +14,6 @@
  ***************************************************************************/
 
 #include "qgsmaptooladdellipse.h"
-#include "qgscompoundcurve.h"
-#include "qgscurvepolygon.h"
 #include "qgsgeometryrubberband.h"
 #include "qgsgeometryutils.h"
 #include "qgslinestring.h"
@@ -25,65 +23,9 @@
 #include "qgssnapindicator.h"
 
 QgsMapToolAddEllipse::QgsMapToolAddEllipse( QgsMapToolCapture *parentTool, QgsMapCanvas *canvas, CaptureMode mode )
-  : QgsMapToolCapture( canvas, QgisApp::instance()->cadDockWidget(), mode )
-  , mParentTool( parentTool )
-  , mSnapIndicator( std::make_unique< QgsSnapIndicator>( canvas ) )
+  : QgsMapToolAddAbstract( parentTool, canvas, mode )
 {
   mToolName = tr( "Add ellipse" );
-
-  clean();
-  connect( QgisApp::instance(), &QgisApp::newProject, this, &QgsMapToolAddEllipse::stopCapturing );
-  connect( QgisApp::instance(), &QgisApp::projectRead, this, &QgsMapToolAddEllipse::stopCapturing );
-}
-
-QgsMapToolAddEllipse::~QgsMapToolAddEllipse()
-{
-  clean();
-}
-
-void QgsMapToolAddEllipse::keyPressEvent( QKeyEvent *e )
-{
-  if ( e && e->isAutoRepeat() )
-  {
-    return;
-  }
-
-  if ( e && e->key() == Qt::Key_Escape )
-  {
-    clean();
-    if ( mParentTool )
-      mParentTool->keyPressEvent( e );
-  }
-
-  if ( e && e->key() == Qt::Key_Backspace )
-  {
-    if ( mPoints.size() == 1 )
-    {
-
-      if ( mTempRubberBand )
-      {
-        delete mTempRubberBand;
-        mTempRubberBand = nullptr;
-      }
-
-      mPoints.clear();
-    }
-    else if ( mPoints.size() > 1 )
-    {
-      mPoints.removeLast();
-
-    }
-    if ( mParentTool )
-      mParentTool->keyPressEvent( e );
-  }
-}
-
-void QgsMapToolAddEllipse::keyReleaseEvent( QKeyEvent *e )
-{
-  if ( e && e->isAutoRepeat() )
-  {
-    return;
-  }
 }
 
 void QgsMapToolAddEllipse::deactivate()
@@ -114,40 +56,8 @@ void QgsMapToolAddEllipse::deactivate()
   QgsMapToolCapture::deactivate();
 }
 
-void QgsMapToolAddEllipse::activate()
-{
-  clean();
-  QgsMapToolCapture::activate();
-}
-
 void QgsMapToolAddEllipse::clean()
 {
-  if ( mTempRubberBand )
-  {
-    delete mTempRubberBand;
-    mTempRubberBand = nullptr;
-  }
-
-  mPoints.clear();
-
-  if ( mParentTool )
-  {
-    mParentTool->deleteTempRubberBand();
-  }
-
+  QgsMapToolAddAbstract::clean();
   mEllipse = QgsEllipse();
-
-  QgsVectorLayer *vLayer = static_cast<QgsVectorLayer *>( QgisApp::instance()->activeLayer() );
-  if ( vLayer )
-    mLayerType = vLayer->geometryType();
-}
-
-void QgsMapToolAddEllipse::release( QgsMapMouseEvent *e )
-{
-  deactivate();
-  if ( mParentTool )
-  {
-    mParentTool->canvasReleaseEvent( e );
-  }
-  activate();
 }
