+++ src/app/qgsmaptooladdellipse.h
@@ -16,7 +16,7 @@
 #ifndef QGSMAPTOOLADDELLIPSE_H
 #define QGSMAPTOOLADDELLIPSE_H
 
-#include "qgsmaptoolcapture.h"
+#include "qgsmaptooladdabstract.h"
 #include "qgsellipse.h"
 #include "qgssettingsregistrycore.h"
 #include "qgis_app.h"
@@ -24,46 +24,23 @@
 class QgsGeometryRubberBand;
 class QgsSnapIndicator;
 
-class APP_EXPORT QgsMapToolAddEllipse: public QgsMapToolCapture
+class APP_EXPORT QgsMapToolAddEllipse: public QgsMapToolAddAbstract
 {
     Q_OBJECT
   public:
     QgsMapToolAddEllipse( QgsMapToolCapture *parentTool, QgsMapCanvas *canvas, CaptureMode mode = CaptureLine );
-    ~QgsMapToolAddEllipse() override;
-
-    void keyPressEvent( QKeyEvent *e ) override;
-    void keyReleaseEvent( QKeyEvent *e ) override;
 
     void deactivate() override;
-
-    void activate() override;
     void clean() override;
 
   protected:
     explicit QgsMapToolAddEllipse( QgsMapCanvas *canvas ) = delete; //forbidden
 
-    //! Convenient method to release (activate/deactivate) tools
-    void release( QgsMapMouseEvent *e );
-
-    /**
-     * The parent map tool, e.g. the add feature tool.
-     * Completed ellipse will be added to this tool by calling its toLineString() method.
-     */
-    QPointer<QgsMapToolCapture> mParentTool;
-    //! Ellipse points (in map coordinates)
-    QgsPointSequence mPoints;
-    //! The rubberband to show the ellipse currently working on
-    QgsGeometryRubberBand *mTempRubberBand = nullptr;
     //! Ellipse
     QgsEllipse mEllipse;
+
     //! convenient method to return the number of segments
     unsigned int segments( ) { return QgsSettingsRegistryCore::settingsDigitizingOffsetQuadSeg.value() * 12; }
-    //! Layer type which will be used for rubberband
-    QgsWkbTypes::GeometryType mLayerType = QgsWkbTypes::LineGeometry;
-
-    //! Snapping indicators
-    std::unique_ptr<QgsSnapIndicator> mSnapIndicator;
-
 };
 
 #endif // QGSMAPTOOLADDELLIPSE_H
