+++ src/app/qgsmaptooladdrectangle.cpp
@@ -26,65 +26,9 @@
 #include "qgssnapindicator.h"
 
 QgsMapToolAddRectangle::QgsMapToolAddRectangle( QgsMapToolCapture *parentTool, QgsMapCanvas *canvas, CaptureMode mode )
-  : QgsMapToolCapture( canvas, QgisApp::instance()->cadDockWidget(), mode )
-  , mParentTool( parentTool )
-  , mSnapIndicator( std::make_unique< QgsSnapIndicator>( canvas ) )
+  : QgsMapToolAddAbstract( parentTool, canvas, mode )
 {
   mToolName = tr( "Add rectangle" );
-
-  clean();
-  connect( QgisApp::instance(), &QgisApp::newProject, this, &QgsMapToolAddRectangle::stopCapturing );
-  connect( QgisApp::instance(), &QgisApp::projectRead, this, &QgsMapToolAddRectangle::stopCapturing );
-}
-
-QgsMapToolAddRectangle::~QgsMapToolAddRectangle()
-{
-  clean();
-}
-
-void QgsMapToolAddRectangle::keyPressEvent( QKeyEvent *e )
-{
-  if ( e && e->isAutoRepeat() )
-  {
-    return;
-  }
-
-  if ( e && e->key() == Qt::Key_Escape )
-  {
-    clean();
-    if ( mParentTool )
-      mParentTool->keyPressEvent( e );
-  }
-
-  if ( e && e->key() == Qt::Key_Backspace )
-  {
-    if ( mPoints.size() == 1 )
-    {
-
-      if ( mTempRubberBand )
-      {
-        delete mTempRubberBand;
-        mTempRubberBand = nullptr;
-      }
-
-      mPoints.clear();
-    }
-    else if ( mPoints.size() > 1 )
-    {
-      mPoints.removeLast();
-
-    }
-    if ( mParentTool )
-      mParentTool->keyPressEvent( e );
-  }
-}
-
-void QgsMapToolAddRectangle::keyReleaseEvent( QKeyEvent *e )
-{
-  if ( e && e->isAutoRepeat() )
-  {
-    return;
-  }
 }
 
 void QgsMapToolAddRectangle::deactivate( )
@@ -115,40 +59,8 @@ void QgsMapToolAddRectangle::deactivate( )
   QgsMapToolCapture::deactivate();
 }
 
-void QgsMapToolAddRectangle::activate()
-{
-  clean();
-  QgsMapToolCapture::activate();
-}
-
 void QgsMapToolAddRectangle::clean()
 {
-  if ( mTempRubberBand )
-  {
-    delete mTempRubberBand;
-    mTempRubberBand = nullptr;
-  }
-
-  mPoints.clear();
-
-  if ( mParentTool )
-  {
-    mParentTool->deleteTempRubberBand();
-  }
-
+  QgsMapToolAddAbstract::clean();
   mRectangle = QgsQuadrilateral();
-
-  QgsVectorLayer *vLayer = static_cast<QgsVectorLayer *>( QgisApp::instance()->activeLayer() );
-  if ( vLayer )
-    mLayerType = vLayer->geometryType();
-}
-
-void QgsMapToolAddRectangle::release( QgsMapMouseEvent *e )
-{
-  deactivate();
-  if ( mParentTool )
-  {
-    mParentTool->canvasReleaseEvent( e );
-  }
-  activate();
 }
