+++ src/app/qgsmaptooladdrectangle.h
@@ -16,53 +16,26 @@
 #ifndef QGSMAPTOOLADDRECTANGLE_H
 #define QGSMAPTOOLADDRECTANGLE_H
 
+#include "qgsmaptooladdabstract.h"
 #include "qgspolygon.h"
-#include "qgsmaptoolcapture.h"
 #include "qgsquadrilateral.h"
 #include "qgis_app.h"
 
-class QgsPolygon;
-class QgsSnapIndicator;
-
-class APP_EXPORT QgsMapToolAddRectangle: public QgsMapToolCapture
+class APP_EXPORT QgsMapToolAddRectangle: public QgsMapToolAddAbstract
 {
     Q_OBJECT
 
   public:
     QgsMapToolAddRectangle( QgsMapToolCapture *parentTool, QgsMapCanvas *canvas, CaptureMode mode = CaptureLine );
-    ~QgsMapToolAddRectangle() override;
-
-    void keyPressEvent( QKeyEvent *e ) override;
-    void keyReleaseEvent( QKeyEvent *e ) override;
 
     void deactivate( ) override;
-
-    void activate() override;
     void clean() override;
 
   protected:
     explicit QgsMapToolAddRectangle( QgsMapCanvas *canvas ) = delete; //forbidden
 
-    //! Convenient method to release (activate/deactivate) tools
-    void release( QgsMapMouseEvent *e );
-
-    /**
-     * The parent map tool, e.g. the add feature tool.
-     * Completed regular shape will be added to this tool by calling its addCurve() method.
-     */
-    QPointer<QgsMapToolCapture> mParentTool;
-    //! Regular Shape points (in map coordinates)
-    QgsPointSequence mPoints;
-    //! The rubberband to show the rectangle currently working on
-    QgsGeometryRubberBand *mTempRubberBand = nullptr;
     //! Rectangle
     QgsQuadrilateral mRectangle;
-
-    //! Layer type which will be used for rubberband
-    QgsWkbTypes::GeometryType mLayerType = QgsWkbTypes::LineGeometry;
-
-    //! Snapping indicators
-    std::unique_ptr<QgsSnapIndicator> mSnapIndicator;
 };
 
 #endif // QGSMAPTOOLADDRECTANGLE_H
