+++ src/app/qgsmaptooladdregularpolygon.cpp
@@ -14,8 +14,6 @@
  ***************************************************************************/
 
 #include "qgsmaptooladdregularpolygon.h"
-#include "qgscompoundcurve.h"
-#include "qgscurvepolygon.h"
 #include "qgsgeometryrubberband.h"
 #include "qgsgeometryutils.h"
 #include "qgsmapcanvas.h"
@@ -25,20 +23,9 @@
 #include "qgssnapindicator.h"
 
 QgsMapToolAddRegularPolygon::QgsMapToolAddRegularPolygon( QgsMapToolCapture *parentTool, QgsMapCanvas *canvas, CaptureMode mode )
-  : QgsMapToolCapture( canvas, QgisApp::instance()->cadDockWidget(), mode )
-  , mParentTool( parentTool )
-  , mSnapIndicator( std::make_unique< QgsSnapIndicator>( canvas ) )
+  : QgsMapToolAddAbstract( parentTool, canvas, mode )
 {
   mToolName = tr( "Add regular polygon" );
-
-  clean();
-  connect( QgisApp::instance(), &QgisApp::newProject, this, &QgsMapToolAddRegularPolygon::stopCapturing );
-  connect( QgisApp::instance(), &QgisApp::projectRead, this, &QgsMapToolAddRegularPolygon::stopCapturing );
-}
-
-QgsMapToolAddRegularPolygon::~QgsMapToolAddRegularPolygon()
-{
-  clean();
 }
 
 void QgsMapToolAddRegularPolygon::createNumberSidesSpinBox()
@@ -57,51 +44,6 @@ void QgsMapToolAddRegularPolygon::deleteNumberSidesSpinBox()
   if ( mNumberSidesSpinBox )
   {
     mNumberSidesSpinBox.reset( nullptr );
-  }
-}
-
-void QgsMapToolAddRegularPolygon::keyPressEvent( QKeyEvent *e )
-{
-  if ( e && e->isAutoRepeat() )
-  {
-    return;
-  }
-
-  if ( e && e->key() == Qt::Key_Escape )
-  {
-    clean();
-    if ( mParentTool )
-      mParentTool->keyPressEvent( e );
-  }
-
-  if ( e && e->key() == Qt::Key_Backspace )
-  {
-    if ( mPoints.size() == 1 )
-    {
-
-      if ( mTempRubberBand )
-      {
-        delete mTempRubberBand;
-        mTempRubberBand = nullptr;
-      }
-
-      mPoints.clear();
-    }
-    else if ( mPoints.size() > 1 )
-    {
-      mPoints.removeLast();
-
-    }
-    if ( mParentTool )
-      mParentTool->keyPressEvent( e );
-  }
-}
-
-void QgsMapToolAddRegularPolygon::keyReleaseEvent( QKeyEvent *e )
-{
-  if ( e && e->isAutoRepeat() )
-  {
-    return;
   }
 }
 
@@ -132,26 +74,9 @@ void QgsMapToolAddRegularPolygon::deactivate()
   QgsMapToolCapture::deactivate();
 }
 
-void QgsMapToolAddRegularPolygon::activate()
-{
-  clean();
-  QgsMapToolCapture::activate();
-}
-
 void QgsMapToolAddRegularPolygon::clean()
 {
-  if ( mTempRubberBand )
-  {
-    delete mTempRubberBand;
-    mTempRubberBand = nullptr;
-  }
-
-  mPoints.clear();
-
-  if ( mParentTool )
-  {
-    mParentTool->deleteTempRubberBand();
-  }
+  QgsMapToolAddAbstract::clean();
 
   if ( mNumberSidesSpinBox )
   {
@@ -159,19 +84,4 @@ void QgsMapToolAddRegularPolygon::clean()
   }
 
   mRegularPolygon = QgsRegularPolygon();
-
-
-  QgsVectorLayer *vLayer = static_cast<QgsVectorLayer *>( QgisApp::instance()->activeLayer() );
-  if ( vLayer )
-    mLayerType = vLayer->geometryType();
-}
-
-void QgsMapToolAddRegularPolygon::release( QgsMapMouseEvent *e )
-{
-  deactivate();
-  if ( mParentTool )
-  {
-    mParentTool->canvasReleaseEvent( e );
-  }
-  activate();
 }
