+++ src/app/qgsmaptooladdregularpolygon.h
@@ -16,28 +16,21 @@
 #ifndef QGSMAPTOOLADDREGULARPOLYGON_H
 #define QGSMAPTOOLADDREGULARPOLYGON_H
 
+#include "qgsmaptooladdabstract.h"
 #include "qgsregularpolygon.h"
-#include "qgsmaptoolcapture.h"
 #include "qgsspinbox.h"
 #include "qgis_app.h"
 
 class QSpinBox;
-class QgsSnapIndicator;
 
-class APP_EXPORT QgsMapToolAddRegularPolygon: public QgsMapToolCapture
+class APP_EXPORT QgsMapToolAddRegularPolygon: public QgsMapToolAddAbstract
 {
     Q_OBJECT
 
   public:
     QgsMapToolAddRegularPolygon( QgsMapToolCapture *parentTool, QgsMapCanvas *canvas, CaptureMode mode = CaptureLine );
-    ~QgsMapToolAddRegularPolygon() override;
-
-    void keyPressEvent( QKeyEvent *e ) override;
-    void keyReleaseEvent( QKeyEvent *e ) override;
 
     void deactivate() override;
-
-    void activate() override;
     void clean() override;
 
   protected:
@@ -51,27 +44,8 @@ class APP_EXPORT QgsMapToolAddRegularPolygon: public QgsMapToolCapture
     //! delete the spin box to enter the number of sides, if it exists
     void deleteNumberSidesSpinBox();
 
-    //! Convenient method to release (activate/deactivate) tools
-    void release( QgsMapMouseEvent *e );
-
-    /**
-     * The parent map tool, e.g. the add feature tool.
-     * Completed regular polygon will be added to this tool by calling its addCurve() method.
-     */
-    QPointer<QgsMapToolCapture> mParentTool;
-    //! Regular Shape points (in map coordinates)
-    QgsPointSequence mPoints;
-    //! The rubberband to show the regular polygon currently working on
-    QgsGeometryRubberBand *mTempRubberBand = nullptr;
     //! Regular shape as a regular polygon
     QgsRegularPolygon mRegularPolygon;
-
-    //! Layer type which will be used for rubberband
-    QgsWkbTypes::GeometryType mLayerType = QgsWkbTypes::LineGeometry;
-
-    //! Snapping indicators
-    std::unique_ptr<QgsSnapIndicator> mSnapIndicator;
-
 };
 
 #endif // QGSMAPTOOLADDREGULARPOLYGON_H
